balance data type change

+++ exporters/jdbc/schema/coin_balances.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from sqlalchemy import Column, PrimaryKeyConstraint
-from sqlalchemy.dialects.postgresql import BYTEA, BIGINT, TIMESTAMP
+from sqlalchemy.dialects.postgresql import BYTEA, BIGINT, TIMESTAMP, NUMERIC
 from exporters.jdbc.schema import Base
 
 
@@ -8,7 +8,7 @@ class CoinBalances(Base):
     __tablename__ = 'address_coin_balances'
 
     address = Column(BYTEA, primary_key=True)
-    balance = Column(BIGINT)
+    balance = Column(NUMERIC(100))
     block_number = Column(BIGINT, primary_key=True)
     block_timestamp = Column(TIMESTAMP)
